+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Fix Semaphore
 + Upgrade dependencies
++ Add `bank-account-transactions` api, reducer, page
 
 ## 0.2.0
 + Add CircleCI